First version of CUDA based code for FPUT system
Some stuff needs to be polished. Energy calculation needs to be implemented. Not all the data is written to file and empty files are opened in those cases. Memory transfer is slow and calculations work with global memory. This must be updated to include shared (update steps can be performed using data in shared memory) and texture memories (using stencils for calculation of nearest neighbor forces). Possibly constant memory to access parameters.